+++ tests/Functions/IteratorMapTest.php
@@ -1,6 +1,7 @@
 <?php
 
 use DoekeNorg\IteratorFunctions\Iterator\MapIterator;
+use DoekeNorg\IteratorFunctions\Tests\Assets\Functions;
 
 /**
  * Tests for {@see iterator_map()}.
@@ -15,6 +16,14 @@ it('maps an iterator', function () {
 
 it('maps an array as an iterator', function () {
     $map_iterator = iterator_map('strtoupper', ['one', 'two', 'three']);
+
+    expect($map_iterator)->toBeInstanceOf(MapIterator::class);
+    expect(iterator_to_array($map_iterator))->toBe(['ONE', 'TWO', 'THREE']);
+});
+
+it('maps with a callback via array params', function () {
+    $functions = new Functions();
+    $map_iterator = iterator_map([$functions, 'strtoupper'], ['one', 'two', 'three']);
 
     expect($map_iterator)->toBeInstanceOf(MapIterator::class);
     expect(iterator_to_array($map_iterator))->toBe(['ONE', 'TWO', 'THREE']);