+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added all `_(u)diff_`, `_(u)assoc` and `_(u)intersect` methods as iterator functions.
 
 ## [1.0.1] - 2021-08-29
 ### Changed